# perf: use lodash to efficiently remove forged transactions