--- a/packages/core-transaction-pool/package.json
+++ b/packages/core-transaction-pool/package.json
@@ -40,6 +40,7 @@
         "delay": "^4.3.0",
         "fs-extra": "^8.1.0",
         "lodash.clonedeep": "^4.5.0",
+        "lodash.differencewith": "^4.5.0",
         "pluralize": "^8.0.0"
     },
     "devDependencies": {